+++ src/data_objects/Media.h
@@ -22,9 +22,9 @@
     NSData *data; 
 }
 
-@property (readonly, assign) int media_id;
-@property (readonly, assign) int game_id;
-@property (readonly, assign) int user_id;
+@property (readonly, assign) long media_id;
+@property (readonly, assign) long game_id;
+@property (readonly, assign) long user_id;
 @property (nonatomic, strong, readonly) NSURL *localURL;
 @property (nonatomic, strong) NSURL *remoteURL;
 @property (nonatomic, strong) NSData *data;
