+++ src/data_objects/Overlay.h
@@ -13,15 +13,15 @@
 //which are derived on the fly from the raw data
 @interface Overlay : NSObject <MKOverlay>
 {
-  int overlay_id;
-  int media_id;
+  long overlay_id;
+  long media_id;
   CLLocation *top_left_corner;
   CLLocation *top_right_corner;
   CLLocation *bottom_left_corner;
 }
 
-@property (nonatomic, assign) int overlay_id;
-@property (nonatomic, assign) int media_id;
+@property (nonatomic, assign) long overlay_id;
+@property (nonatomic, assign) long media_id;
 @property (nonatomic, strong) CLLocation *top_left_corner;
 @property (nonatomic, strong) CLLocation *top_right_corner;
 @property (nonatomic, strong) CLLocation *bottom_left_corner;
