+++ src/data_objects/Plaque.h
@@ -11,20 +11,20 @@
 
 @interface Plaque : NSObject <InstantiableProtocol>
 {
-  int plaque_id;
+  long plaque_id;
   NSString *name;
   NSString *desc;
-  int icon_media_id;
-  int media_id;
-  int event_package_id;
+  long icon_media_id;
+  long media_id;
+  long event_package_id;
 }
 
-@property(nonatomic, assign) int plaque_id;
+@property(nonatomic, assign) long plaque_id;
 @property(nonatomic, strong) NSString *name;
 @property(nonatomic, strong) NSString *desc;
-@property(nonatomic, assign) int icon_media_id;
-@property(nonatomic, assign) int media_id;
-@property(nonatomic, assign) int event_package_id;
+@property(nonatomic, assign) long icon_media_id;
+@property(nonatomic, assign) long media_id;
+@property(nonatomic, assign) long event_package_id;
 
 - (id) initWithDictionary:(NSDictionary *)dict;
