+++ src/data_objects/Scene.m
@@ -35,7 +35,7 @@
 }
 
 //To comply w/ instantiable protocol. should get default image later.
-- (int) icon_media_id
+- (long) icon_media_id
 {
     return 0;
 }
