+++ src/models/dialogs/DialogsModel.h
@@ -14,12 +14,12 @@
 
 @interface DialogsModel : NSObject
 
-- (Dialog *) dialogForId:(int)dialog_id;
-- (DialogCharacter *) characterForId:(int)dialog_character_id;
-- (DialogScript *) scriptForId:(int)dialog_script_id;
-- (DialogOption *) optionForId:(int)dialog_option_id;
+- (Dialog *) dialogForId:(long)dialog_id;
+- (DialogCharacter *) characterForId:(long)dialog_character_id;
+- (DialogScript *) scriptForId:(long)dialog_script_id;
+- (DialogOption *) optionForId:(long)dialog_option_id;
 - (void) requestDialogs;
 - (void) clearGameData;
-- (void) requestPlayerOptionsForDialogId:(int)dialog_id scriptId:(int)dialog_script_id;
+- (void) requestPlayerOptionsForDialogId:(long)dialog_id scriptId:(long)dialog_script_id;
 
 @end
