+++ src/models/notes/NotesModel.h
@@ -18,10 +18,10 @@
 {
 }
 
-- (Note *) noteForId:(int)note_id;
+- (Note *) noteForId:(long)note_id;
 - (void) requestNotes;
 
-- (NoteComment *) noteCommentForId:(int)note_comment_id;
+- (NoteComment *) noteCommentForId:(long)note_comment_id;
 - (void) requestNoteComments;
 
 - (void) clearGameData;
@@ -29,17 +29,17 @@
 
 - (void) createNote:(Note *)n withTag:(Tag *)t media:(Media *)m trigger:(Trigger *)tr;
 - (void) saveNote:(Note *)n withTag:(Tag *)t media:(Media *)m trigger:(Trigger *)tr;
-- (void) deleteNoteId:(int)note_id;
+- (void) deleteNoteId:(long)note_id;
 
 - (void) createNoteComment:(NoteComment *)n;
 - (void) saveNoteComment:(NoteComment *)n;
-- (void) deleteNoteCommentId:(int)note_comment_id;
+- (void) deleteNoteCommentId:(long)note_comment_id;
 
 - (NSArray *) notes;
 - (NSArray *) playerNotes;
 - (NSArray *) listNotes;
 - (NSArray *) notesMatchingTag:(Tag *)tag;
 - (NSArray *) noteComments;
-- (NSArray *) noteCommentsForNoteId:(int)note_id;
+- (NSArray *) noteCommentsForNoteId:(long)note_id;
 
 @end
