+++ src/object_controllers/item/ItemActionViewController.h
@@ -14,8 +14,8 @@
 @class Item;
 
 @protocol ItemActionViewControllerDelegate
-- (void) amtChosen:(int)amt positive:(BOOL)p;
+- (void) amtChosen:(long)amt positive:(BOOL)p;
 @end
 @interface ItemActionViewController : ARISViewController
-- (id) initWithPrompt:(NSString *)s positive:(BOOL)p maxqty:(int)q delegate:(id)d;
+- (id) initWithPrompt:(NSString *)s positive:(BOOL)p maxqty:(long)q delegate:(id)d;
 @end
